refactor: rename Login folder to AuthLayout
