+++ src/components/index.ts
@@ -10,7 +10,7 @@ export { FavoritesList } from './FavoritesList/FavoritesList'
 export { IconButton } from './IconButton/IconButton'
 export { Info } from './Info/Info'
 export { LoadingError } from './LoadingError/LoadingError'
-export { Login } from './Login/Login'
+export { Login } from './AuthLayout/Login'
 export { MainLayout } from './MainLayout/MainLayout'
 export { ProtectedRoute } from './ProtectedRoute/ProtectedRoute'
 export { SelectField } from './SelectField/SelectField'
